+++ evals/elsuite/utils.py
@@ -13,6 +13,11 @@ from evals.prompt.base import (
     is_chat_prompt,
 )
 
+from langchain.text_splitter import RecursiveCharacterTextSplitter
+from langchain.embeddings.openai import OpenAIEmbeddings
+from langchain.docstore.document import Document
+from langchain.vectorstores import FAISS
+
 
 def get_answer(text, answer_prompt, ignore_case=False):
     if ignore_case:
@@ -141,6 +146,49 @@ def format_prompt(
     return prompt
 
 
+def process_text(text):
+    docs = text_to_docs(text)
+    embeddings = OpenAIEmbeddings()
+    db = FAISS.from_documents(docs, embeddings)
+    return db
+
+def retrieve_context_from_db(query, db, k=3):
+    documents = db.similarity_search(query=query, k=3)
+    return ''.join([doc.page_content for doc in documents])
+
+def text_to_docs(text):
+    """Converts a string or list of strings to a list of Documents
+    with metadata."""
+    if isinstance(text, str):
+        # Take a single string as one page
+        text = [text]
+    page_docs = [Document(page_content=page) for page in text]
+
+    # Add page numbers as metadata
+    for i, doc in enumerate(page_docs):
+        doc.metadata["page"] = i + 1
+
+    # Split pages into chunks
+    doc_chunks = []
+
+    for doc in page_docs:
+        text_splitter = RecursiveCharacterTextSplitter(
+            chunk_size=2000,
+            separators=["\n\n", "\n", ".", "!", "?", ",", " ", ""],
+            chunk_overlap=0,
+        )
+        chunks = text_splitter.split_text(doc.page_content)
+        for i, chunk in enumerate(chunks):
+            doc = Document(
+                page_content=chunk, metadata={"page": doc.metadata["page"], "chunk": i}
+            )
+            # Add sources a metadata
+            doc.metadata["source"] = f"{doc.metadata['page']}-{doc.metadata['chunk']}"
+            doc_chunks.append(doc)
+    print("Length of doc_chunks: ", len(doc_chunks))
+    return doc_chunks
+
+
 class PromptFn:
     """
     Wrap calls to a completion_fn with a prompt template with applicable keyword args.
@@ -179,7 +227,7 @@ class PromptFn:
         else:
             # Prompt is a string
             prompt = format_necessary(self.prompt, **kwargs)
-
+        #print("The final prompt is", prompt)
         result = self.completion_fn(
             prompt=prompt,
             max_tokens=self.max_tokens,
@@ -190,5 +238,6 @@ class PromptFn:
             n=(1 if self.n_samples is None else self.n_samples),
             **self.completion_kwargs,
         )
+        #print("result is", result.get_completions()[0])
         sampled = result.get_completions()[0]
         return sampled, prompt
